Move /chat, /chats and /message code to separate scripts
* Move code from server.js to...
  + backend/routes/chat.js
  + backend/routes/getChats.js
  + backend/routes/message.js
  ... and edit as necessary
* server.js
  + Add import statements for the above scripts
  + Comment out import statements that are no longer needed
  + Comment out code that has been moved to the above scripts
  + Edit /chat, /chats and /message routes to use imported elements
