+++ backend/package.json
@@ -17,7 +17,6 @@
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^6.3.3",
     "multer": "^1.4.5-lts.1",
-    "node": "^17.7.2",
     "node-email-validation": "^1.0.4",
     "nodemon": "^2.0.16",
     "validator": "^13.7.0"
